CEXT-6507: split webhooks helpers.ts into types/utils/plan/apply
helpers.ts had grown to hold validation, install/uninstall, planning,
and apply logic all in one file. Split it to mirror the events domain's
structure (types.ts, utils.ts, plan.ts, apply.ts), leaving helpers.ts
with just validate/install/uninstall. Tests split along the same lines.

Also trimmed several oversized comments down to the essential why.
